requirements.txt
aws-cdk-lib==2.74.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-apigatewayv2-alpha
aws-cdk.aws_apigatewayv2_integrations_alpha
aws_cdk.aws_apigatewayv2_authorizers_alpha
boto3
requests